+++ DomainManagement/changelog.md
@@ -2,6 +2,8 @@
 
 ## ???
 
+- New: AccessRuleMode Component - allows controlling, how individual AccessRules are being processed.
+- Upd: AccessRule - added support for AccessRule processing modes defined by the AccessRuleMode Component.
 - Fix: Component Group - Renaming groups & delta between Name and SamAccountName will now be properly detected and handled.
 - Fix: Invoke-DMUser ignores name updates